+++ packages/server/lib/automation/cookies.ts
@@ -5,8 +5,9 @@ import { isHostOnlyCookie } from '../browsers/cdp_automation'
 
 export interface AutomationCookie {
   domain: string
-  expiry: number | null
+  expiry: 'Infinity' | '-Infinity' | number | null
   httpOnly: boolean
+  hostOnly: boolean
   maxAge: 'Infinity' | '-Infinity' | number | null
   name: string
   path: string | null
@@ -31,9 +32,20 @@ const normalizeCookieProps = function (props) {
     return props
   }
 
+  // if the cookie is stored inside the server side cookie jar,
+  // we want to make the automation client aware so the domain property
+  // isn't mutated to prevent duplicate setting of cookies from different contexts.
+  // This should be handled by the hostOnly property
+
   const cookie = _.pick(props, COOKIE_PROPERTIES)
 
-  if (props.expiry != null) {
+  if (props.expiry === '-Infinity') {
+    cookie.expiry = -Infinity
+    // set the cookie to expired so when set, the cookie is removed
+    cookie.expirationDate = 0
+  } else if (props.expiry === 'Infinity') {
+    cookie.expiry = null
+  } else if (props.expiry != null) {
     // when sending cookie props we need to convert
     // expiry to expirationDate
     delete cookie.expiry