+++ system-tests/projects/coffee-react-interop/lib/dom.jsx
@@ -1,5 +1,5 @@
-import React from "react";
+import React from 'react'
 
 <div className="dom" />
 
-export default "dom"
+export default 'dom'